Detect trees from satellite images/requirements.txt
gunicorn
python-dotenv
numpy
requests
fast-dash